refactor: renommage filtres
